Added tests
